+++ src/Service/MediaConvert/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: This release includes support for creating I-frame only video segments for DASH trick play.
+- AWS api-change: This release adds the ability to search for historical job records within the management console using a search box and/or via the SDK/CLI with partial string matching search on input file name.
+
 ### Changed
 
 - Add `Accept: application/json` header in request to fix incompatibility with 3rd party providers